pyproject.toml
[project]
name = "glow"
version = "0.1.0"
description = "GLow - A Flower Based Gossip Learning Strategy"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Aitor Belenguer"}
]
keywords = ["federated-learning", "gossip-learning", "flower", "machine-learning"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pandas>=1.5.0",
    "scikit-learn>=1.2.0",
    "pyyaml>=6.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.7.0",
    "flwr[simulation]==1.25.0",
    "setuptools>=78.1.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools>=78.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "client",
    "cnl_pytorch",
    "dataset",
    "FL_hydra_main",
    "hydra_main",
    "main",
    "model",
    "server",
]

[tool.setuptools.packages.find]
include = ["custom_strategies*", "flwr_lib_modifications*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # allow raise without from in exception handlers
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# TODO: setup this for some basic tests
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing --cov-report=html"
